% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_excel_df}
\alias{clean_excel_df}
\title{Clean Data Frame for Excel Export}
\usage{
clean_excel_df(dat)
}
\arguments{
\item{dat}{Data frame.}
}
\value{
Cleaned data frame.
}
\description{
Applies Excel-safe text cleaning across all columns.
}
