% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_excel_text}
\alias{clean_excel_text}
\title{Clean Character Values for Excel}
\usage{
clean_excel_text(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Cleaned vector.
}
\description{
Applies Excel-safe text cleaning to a vector.
}
