% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsr.R
\name{dsr}
\alias{dsr}
\alias{DSR}
\title{Launch Data Specs Reviewer}
\usage{
dsr(...)

DSR(...)
}
\arguments{
\item{...}{Arguments passed to \code{data_review()}.}
}
\description{
Shortcut for launching the Data Specs Reviewer application.
}
