% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{or_else}
\alias{or_else}
\title{Return a Default Value When Missing}
\usage{
or_else(x, y)
}
\arguments{
\item{x}{Object to evaluate.}

\item{y}{Fallback value returned when \code{x} is NULL,
empty, or entirely missing.}
}
\value{
Either \code{x} or \code{y}.
}
\description{
Returns a fallback value when the supplied object is NULL,
empty, or contains only missing values.
}
\examples{
or_else(NULL, "ABC")
or_else("", "ABC")

}
