% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_one_file}
\alias{read_one_file}
\title{Read Supported Data Files}
\usage{
read_one_file(path, original_name, read_all_sheets = TRUE)
}
\arguments{
\item{path}{File path.}

\item{original_name}{Original file name.}

\item{read_all_sheets}{Logical.
If TRUE, all worksheets in an Excel workbook are imported.
If FALSE, only the first worksheet is imported.}
}
\value{
Named list of datasets.
}
\description{
Imports SAS, Excel, XLSB, and CSV files.
}
